backend/requirements.txt
Django
django-cors-headers
djangorestframework
requests
mysqlclient
tensorflow